Add E2E tests that run the CLI against Insights, relocate cobra code to a new cli pkg,  (#114)
* Relocate cobra code to a new cli pkg, add E2E tests that run insights-cli against Insights

* Relocate vault installation / sourcing environment

* Switch to goreleaser image from CircleCI

* Fix setting E2E env vars, also set them for the second place we run tests

* Fix comment in CI config about all the places the goreleaser image is referenced
