+++ cmd/insights/main.go
@@ -1,4 +1,4 @@
-// Copyright 2020 Fairwinds
+// Copyright 2020 FairwindsOps Inc
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -10,26 +10,16 @@
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
-// limitations under the License
+// limitations under the License.
 
 package main
 
 import (
-	"fmt"
-	cliversion "github.com/fairwindsops/insights-cli/pkg/version"
+	"os"
 
-	"github.com/spf13/cobra"
+	"github.com/fairwindsops/insights-cli/pkg/cli"
 )
 
-func init() {
-	rootCmd.AddCommand(versionCmd)
-}
-
-var versionCmd = &cobra.Command{
-	Use:   "version",
-	Short: "Prints the current version of the tool.",
-	Long:  `Prints the current version.`,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println(cliversion.String())
-	},
+func main() {
+	os.Exit(cli.Run())
 }
