+++ pkg/cli/version.go
@@ -1,0 +1,35 @@
+// Copyright 2020 Fairwinds
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License
+
+package cli
+
+import (
+	"fmt"
+	cliversion "github.com/fairwindsops/insights-cli/pkg/version"
+
+	"github.com/spf13/cobra"
+)
+
+func init() {
+	rootCmd.AddCommand(versionCmd)
+}
+
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "Prints the current version of the tool.",
+	Long:  `Prints the current version.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println(cliversion.String())
+	},
+}
