+++ testdata/push_all_want_output.txt
@@ -1,0 +1,8 @@
+.
+   opa
+       v1test (v1)
+           deployments
+       v2test (v2)
+   rules
+       pluto severity increase
+
